refactor(pkg/devcontainer): split functions across file to reduce file size
